feat(core): Add v2 parallel workflow execution with performance validation
- Add v2 execution mode for parallel workflow execution
- Implement data cloning for context isolation between parallel nodes
- Add maxParallel workflow setting with UI support
- Include performance test showing 2.8x improvement over sequential execution
- Maintain 100% backward compatibility with existing v0/v1 modes
